Remove uncessary Factory object
