+++ activerecord/lib/active_record/query_logs.rb
@@ -91,8 +91,15 @@ module ActiveRecord
       end
 
       # Updates the formatter to be what the passed in format is.
-      def update_formatter(format = :legacy)
-        self.tags_formatter = QueryLogs::FormatterFactory.from_symbol(format)
+      def update_formatter(format)
+        self.tags_formatter = case format
+        when :legacy
+          Formatter.new(key_value_separator: ":")
+        when :sqlcommenter
+          QuotingFormatter.new(key_value_separator: "=")
+        else
+          raise ArgumentError, "Formatter is unsupported: #{formatter}"
+        end
       end
 
       ActiveSupport::ExecutionContext.after_change { ActiveRecord::QueryLogs.clear_cache }
@@ -109,7 +116,7 @@ module ActiveRecord
         end
 
         def formatter
-          self.tags_formatter ||= QueryLogs::FormatterFactory.from_symbol(:legacy)
+          self.tags_formatter || self.update_formatter(:legacy)
         end
 
         def uncached_comment
